chore(subscribed-services): test refactor, app/domain layer distinction, ternary changes from pr review
